Module Config.TYPES

type options = {

timeout : float;

only_new_tests : bool;

keep_error_cases : bool;

search : string list;

label_handling : label_handling;

replay : replay;

replay_max_concurrency : int;

libc : string;

}

and replay =

| Inhibit
(*
do not replay any test
*)

| Enable
(*
on-the-fly replay if relevant (w.r.t label handling policy)
*)

| Delayed
(*
delay any replay until after the main klee process terminates
*)

| Force
(*
force on-the-fly replay, whatever the label handling policy
*)

and label_handling =

| Ignore
(*
pc_label(e, id) amounts to no C code
*)

| Naive
(*
pc_label(e, id) amounts to if e set_cover(id)
*)

| Optimized of {

benign_test_policy : benign_test_policy;

}

and benign_test_policy =

| Keep
(*
generate and keep all tests
*)

| Reject
(*
do not keep any test that does not reach a label
*)

| Avoid
(*
do not generate any test that neither reaches a label nor triggers a runtime error
*)

how are tests that pass no label treated?