Parameter Make_io.Test_repr

module Params : sig ... end

Encoding parameters for the representation

val size_bounds : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> int * int

size_bounds typ returns a pair of the minimal structure size (i.e, assuming all pointers are NULL), and maximal size (i.e., the total memory footprint with pointers unfolded according to Encoding.max_ptr_array_length)

val reachable_types_with_pointers : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> Sc_C.Type_collections.SET.t

val fold_access_paths : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> ?max_depth:int -> ?f_prm: (Sc_C.Types.access_path_suffix option -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (Sc_C.Types.access_path_suffix option -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

fold_access_paths typ ~max_depth ~f_prm ~f_ptr acc folds f_prm and f_ptr over every canonical access path suffix that may be a valid memory access from a value of type typ (cf Encoding.ctyp_fold_access_paths).

val fold_direct_access_paths : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> ?f_prm: (Sc_C.Types.access_path_suffix option -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (Sc_C.Types.access_path_suffix option -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

fold_direct_access_paths typ ~f_prm ~f_ptr acc folds f_prm and f_ptr over every field defined in type typ; referenced types are NOT visited recursively (cf Encoding.ctyp_fold_direct_access_paths).

val fold_reachable_pointers : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> ?max_depth:int -> f: (Sc_values__.Cvalues.TYPES.ptr_attributes -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> merge:('a NEL.t -> 'a) -> 'a -> 'a

fold_reachable_pointers ctyp ~max_depth ~f ~merge acc folds f over every pointer field defined in type ctyp; pointed types are visited recursively (up to and including max_depth), and results of accumulation over union fields are computed using merge. max_depth defaults to max_non_nil_ptr_depth (given via encoding parameters) (cf Encoding.ctyp_fold_reachable_pointers).

val visit_access_paths : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> ?max_depth:int -> ?f_prm: (Sc_C.Types.access_path_suffix option -> 'a -> 'a Sc_values__.Cvalues.visit_action) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (Sc_C.Types.access_path_suffix option -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a Sc_values__.Cvalues.visit_action) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

visit_access_paths typ ~max_depth ~f_prm ~f_ptr acc visits every canonical access path suffix that may be a valid memory access from a value of type typ using f_prm and f_ptr (cf Encoding.ctyp_visit_access_paths).

val fold_fields : Sc_values__.Cvalues.struct_ Sc_values__.Cvalues.ctyp -> ?f_prm:(string -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> ?f_ptr: (string -> [ `Carray_with_bound_length of int | `Carray_with_length_field of Sc_values__.Cvalues.TYPES.field_access | `Cstring ] -> 'a -> 'a) Sc_values__.Cvalues.TYPES.ctyp_consumer -> 'a -> 'a

fold_fields typ ~f_prm ~f_ptr acc folds f_prm and/or f_ptr over the fields of typ: f_ptr is given the type t for each pointer field of type t*, whereas f_prm is called with type t for every other field of type t.

module Val : sig ... end